refactor: package chart-ui integrated into plugins
