+++ vben-admin.code-workspace
@@ -105,10 +105,6 @@
       "path": "packages/effects/access",
     },
     {
-      "name": "@vben/chart-ui",
-      "path": "packages/effects/chart-ui",
-    },
-    {
       "name": "@vben/common-ui",
       "path": "packages/effects/common-ui",
     },
@@ -119,6 +115,10 @@
     {
       "name": "@vben/layouts",
       "path": "packages/effects/layouts",
+    },
+    {
+      "name": "@vben/plugins",
+      "path": "packages/effects/plugins",
     },
     {
       "name": "@vben/request",
